Refactor database protocols and extract RecordLookup
Moved the RecordLookup protocol and its CKDatabase extension to a new file for better separation of concerns. Updated AppDatabase typealias to use RecordLookup and RecordReader, and removed write methods from DefaultDatabase to clarify its read-only interface.
